ettlex_core::render::bundle_render

Function render_leaf_bundle

pub fn render_leaf_bundle(
    store: &Store,
    leaf_id: &str,
    leaf_ep_ordinal: Option<u32>,
) -> Result<String>

Render a leaf bundle to Markdown

Generates a complete Markdown bundle by computing the EPT from root to leaf, then aggregating the WHY/WHAT/HOW content from all EPs in the path.

The output includes:

Title showing the full RT path (Root > Mid > Leaf)

Aggregated WHY sections from all EPs

Aggregated WHAT sections from all EPs

Aggregated HOW sections from all EPs

Arguments

store - Reference to the Store

leaf_id - ID of the leaf Ettle

leaf_ep_ordinal - Optional ordinal of the leaf EP to end with

Returns

Markdown string representation of the bundle

Errors

EttleNotFound - If leaf doesn’t exist

EptMissingMapping / EptDuplicateMapping - EPT computation errors

EptAmbiguousLeafEp - If leaf has multiple EPs and ordinal not specified